refactor to create common markdown types
